+++ src/EnrolmentWorkflow/Request/Repository.php
@@ -92,7 +92,7 @@ final class Repository
      */
     public function deleteRequests(int $step_id)/*: void*/
     {
-        foreach ($this->getRequests(null, null, $step_id) as $request) {
+        foreach ($this->getRequests(null, $step_id) as $request) {
             $this->deleteRequest($request);
         }
     }
@@ -153,14 +153,13 @@ final class Repository
 
 
     /**
-     * @param int|null    $check_user_id
      * @param int|null    $obj_ref_id
      * @param int|null    $step_id
-     * @param int|null    $user_id
+     * @param array|null  $user_id
      * @param array|null  $responsible_user_ids
      * @param string|null $object_title
      * @param int|null    $workflow_id
-     * @param bool|null   $accepted
+     * @param bool|null   $edited
      * @param string|null $user_lastname
      * @param string|null $user_firstname
      * @param string|null $user_email
@@ -168,9 +167,9 @@ final class Repository
      *
      * @return Request[]
      */
-    public function getRequests(/*?*/ int $check_user_id = null,/*?*/ int $obj_ref_id = null, /*?*/ int $step_id = null,/*?*/ int $user_id = null,/*?*/ array $responsible_user_ids = null, /*?*/
+    public function getRequests( /*?*/ int $obj_ref_id = null, /*?*/ int $step_id = null,/*?*/ array $user_id = null,/*?*/ array $responsible_user_ids = null, /*?*/
         string $object_title = null,/*?*/
-        int $workflow_id = null, /*?*/ bool $accepted = null,/*?*/ string $user_lastname = null,/*?*/ string $user_firstname = null, /*?*/ string $user_email = null, /*?*/
+        int $workflow_id = null, /*?*/ bool $edited = null,/*?*/ string $user_lastname = null,/*?*/ string $user_firstname = null, /*?*/ string $user_email = null, /*?*/
         string $user_org_units = null
     ) : array {
         if (!self::srUserEnrolment()->enrolmentWorkflow()->isEnabled()) {
@@ -195,19 +194,11 @@ final class Repository
             $wheres["responsible_users"] = $responsible_user_ids;
         }
 
-        if ($accepted !== null) {
-            $wheres["accepted"] = $accepted;
+        if ($edited !== null) {
+            $wheres["accepted"] = $edited;
         }
 
         $requests = Request::where($wheres)->get();
-
-        if (!empty($check_user_id)) {
-            $requests = array_filter($requests, function (Request $request) use ($check_user_id): bool {
-                return ($request->getUserId() === $check_user_id || $request->getCreateUserId() === $check_user_id || $request->getAcceptUserId() === $check_user_id
-                    || in_array($check_user_id, $request->getResponsibleUsers())
-                    || $this->userHasReadRole($check_user_id));
-            });
-        }
 
         if (!empty($object_title)) {
             $requests = array_filter($requests, function (Request $request) use ($object_title): bool {
@@ -261,7 +252,7 @@ final class Repository
             $check_user_id = self::dic()->user()->getId();
         }
 
-        return (self::srUserEnrolment()->userHasRole($check_user_id) || self::srUserEnrolment()->enrolmentWorkflow()->requests()->userHasReadRole($check_user_id));
+        return self::srUserEnrolment()->userHasRole($check_user_id);
     }