+++ src/EnrolmentWorkflow/Rule/NoOtherRequests/NoOtherRequestsChecker.php
@@ -34,7 +34,7 @@ class NoOtherRequestsChecker extends AbstractRuleChecker
      */
     public function check(int $user_id, int $obj_ref_id) : bool
     {
-        return (count(self::srUserEnrolment()->enrolmentWorkflow()->requests()->getRequests(null, $obj_ref_id, null, $user_id)) < 2);
+        return (count(self::srUserEnrolment()->enrolmentWorkflow()->requests()->getRequests($obj_ref_id, null, [$user_id])) < 2);
     }