+++ src/EnrolmentWorkflow/Step/StepFormGUI.php
@@ -99,15 +99,15 @@ class StepFormGUI extends PropertyFormGUI
                 ], true),
                 "setTitle"              => $this->txt("action_title")
             ],
-            "action_accept_titles" => [
+            "action_edit_titles" => [
                 self::PROPERTY_CLASS    => TabsInputGUI::class,
                 self::PROPERTY_REQUIRED => true,
                 self::PROPERTY_SUBITEMS => MultilangualTabsInputGUI::generate([
-                    "action_accept_title" => [
+                    "action_edit_title" => [
                         self::PROPERTY_CLASS => ilTextInputGUI::class
                     ]
                 ], true),
-                "setTitle"              => $this->txt("action_accept_title")
+                "setTitle"              => $this->txt("action_edit_title")
             ]
         ];
     }