+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [x]
 - 'Delete requests' action
 - Roles for read requests
+- Assign responsible users by global role
+- Requests tabs
+- Fixes
 
 ## [3.7.0]
 - Members view